+++ fsleyes_widgets/tests/test_colourbarbitmap.py
@@ -25,6 +25,18 @@ from . import compare_images
 datadir = op.join(op.dirname(__file__), 'testdata', 'colourbarbitmap')
 
 
+def gen_file_id(*args):
+    """Generate a "safe" string for use in file names. """
+    fid = '_'.join(map(str, args))
+    for c in '[](){}, !@#$%&*;:=<>|/?\\"\'':
+        fid = fid.replace(c, '_')
+
+    while '__' in fid:         fid = fid.replace('__', '_')
+    while fid.endswith('_'):   fid = fid[:-1]
+    while fid.startswith('_'): fid = fid[1:]
+    return fid
+
+
 def _compare(bmp, fname):
 
     bmp       = bmp.transpose((1, 0, 2))
@@ -47,21 +59,20 @@ def _compare(bmp, fname):
 def test_standard_usage():
 
     cmaps           = ['Greys', 'Blues']
-    sizes           = [(100, 25)]
     cmapResolutions = [6, 256]
     orientations    = ['vertical', 'horizontal']
     alphas          = [0.25, 1.0]
     bgColours       = [(0, 0, 0, 0), (0, 0, 0, 1), (1, 0, 0, 1)]
+    size            = (100, 25)
 
     testcases = it.product(cmaps,
-                           sizes,
                            cmapResolutions,
                            orientations,
                            alphas,
                            bgColours)
 
     for testcase in testcases:
-        cmap, size, cmapRes, orient, alpha, bgColour = testcase
+        cmap, cmapRes, orient, alpha, bgColour = testcase
 
         if orient == 'vertical': height, width = size
         else:                    width, height = size
@@ -74,13 +85,8 @@ def test_standard_usage():
                                       orientation=orient,
                                       bgColour=bgColour)
 
-        fname = '_'.join(map(str, [
-            cmap,
-            cmapRes,
-            orient,
-            alpha] + list(bgColour)))
-
-        fname = 'standard_usage_{}.png'.format(fname)
+        fname = gen_file_id('standard_usage', *testcase)
+        fname = f'{fname}.png'
 
         assert _compare(bmp, fname)
 
@@ -103,8 +109,8 @@ def test_negCmap_invert():
 
     for cmap, negCmap, invert, orient in testcases:
 
-        fname = '_'.join(map(str, [cmap, negCmap, invert, orient]))
-        fname = 'negCmap_invert_{}.png'.format(fname)
+        fname = gen_file_id('negCmap_invert', cmap, negCmap, invert, orient)
+        fname = f'{fname}.png'
 
         if orient == 'vertical': height, width = size
         else:                    width, height = size
@@ -136,8 +142,8 @@ def test_gamma():
         if not useNegCmap:
             negCmap = None
 
-        fname = '_'.join(map(str, [gamma, invert, cmap, negCmap]))
-        fname = 'gamma_{}.png'.format(fname)
+        fname = gen_file_id('gamma', gamma, invert, cmap, negCmap)
+        fname = f'{fname}.png'
 
         bmp = cbarbmp.colourBarBitmap(cmap,
                                       width,
@@ -170,13 +176,9 @@ def test_logScaleRange_interp():
 
     for logScale, interp, res, cmap in testcases:
 
-        if logScale is None:
-            fname = '_'.join(map(str, [logScale, interp, res, cmap]))
-        else:
-            fname = '_'.join(map(str, [logScale[0], logScale[1], interp, res,
-                                       cmap]))
-
-        fname = 'logScaleRange_interp_{}.png'.format(fname)
+        fname = gen_file_id('logScaleRange_interp',
+                            logScale, interp, res, cmap)
+        fname = f'{fname}.png'
 
         bmp = cbarbmp.colourBarBitmap(cmap,
                                       width,
@@ -212,8 +214,8 @@ def test_label():
                                       fontsize=size,
                                       textColour=colour)
 
-        fname = '_'.join(map(str, [orient, side, size] + list(colour)))
-        fname = 'label_{}.png'.format(fname)
+        fname = gen_file_id('label', orient, side, size, colour)
+        fname = f'{fname}.png'
 
         assert _compare(bmp, fname)
 
@@ -250,9 +252,8 @@ def test_ticks():
                                       fontsize=size,
                                       textColour=colour)
 
-        fname = [orient, side, size] + list(colour) + labels
-        fname = '_'.join(map(str, fname))
-        fname = 'ticks_{}.png'.format(fname)
+        fname = gen_file_id('ticks', orient, side, size, colour, labels)
+        fname = f'{fname}.png'
 
         assert _compare(bmp, fname)
 
@@ -289,8 +290,8 @@ def test_tickalign():
 
 
         falign = ''.join([a[0] for a in align])
-        fname  = '_'.join([orient, side, falign])
-        fname  = 'tickalign_{}.png'.format(fname)
+        fname  = gen_file_id('tickalign', orient, side, falign)
+        fname  = f'{fname}.png'
 
         assert _compare(bmp, fname)
 
@@ -325,8 +326,8 @@ def test_label_and_ticks():
                                       fontsize=size,
                                       textColour=(0, 0, 0.2, 1))
 
-        fname = '_'.join([orient, side, str(size)])
-        fname = 'label_and_ticks_{}.png'.format(fname)
+        fname = gen_file_id('label_and_ticks', orient, side, size)
+        fname = f'{fname}.png'
 
         assert _compare(bmp, fname)
 
@@ -360,8 +361,8 @@ def test_scale():
                                       scale=scale,
                                       textColour=(0, 0, 0.2, 1))
 
-        fname = '_'.join([str(scale), orient, side])
-        fname = 'scale_{}.png'.format(fname)
+        fname = gen_file_id('scale', scale, orient, side)
+        fname = f'{fname}.png'
         assert _compare(bmp, fname)
 
 
@@ -375,8 +376,8 @@ def test_negCmap_ticks():
 
     for cmap, negCmap, ticks in testcases:
 
-        fname = '_'.join(map(str, [cmap, negCmap] + ticks))
-        fname = 'negCmap_invert_ticks_{}.png'.format(fname)
+        fname = gen_file_id(cmap, negCmap, ticks)
+        fname = f'negCmap_invert_ticks_{fname}.png'
 
         ticklabels = ['{:0.2f}'.format(t) for t in ticks]
         tickalign = ['left'] + ['center'] * (len(ticks) - 2) + ['right']
@@ -390,6 +391,83 @@ def test_negCmap_ticks():
                                       ticks=ticks,
                                       tickalign=tickalign,
                                       ticklabels=ticklabels)
+        assert _compare(bmp, fname)
+
+
+def test_modAlpha():
+
+    cmaps         = ['Reds']
+    alphas        = [0.5, 1.0]
+    bgColours     = [(0, 0, 0, 0), (0, 0, 0, 1), (1, 0, 0, 1)]
+    dispRanges    = [(-5, 5)]
+    modRanges     = [(-2.5, 2.5), (-5, 5), (-10, -5), (5, 10)]
+    invModAlphas  = [False, True]
+    width, height = 300, 100
+
+
+    testcases = it.product(cmaps,
+                           alphas,
+                           bgColours,
+                           dispRanges,
+                           modRanges,
+                           invModAlphas)
+
+    for testcase in testcases:
+        cmap, alpha, bgColour, dispRange, modRange, invModAlpha = testcase
+
+        bmp = cbarbmp.colourBarBitmap(cmap,
+                                      width,
+                                      height,
+                                      alpha=alpha,
+                                      bgColour=bgColour,
+                                      orientation='horizontal',
+                                      modAlpha=True,
+                                      invModAlpha=invModAlpha,
+                                      displayRange=dispRange,
+                                      modRange=modRange)
+
+        fname = gen_file_id('modAlpha', *testcase)
+        fname = f'{fname}.png'
+
+        assert _compare(bmp, fname)
+
+
+def test_modAlpha_negCmap():
+
+    cmaps         = ['Reds']
+    negCmaps      = ['Blues']
+    alphas        = [0.5, 1.0]
+    bgColours     = [(0, 0, 0, 0), (0, 0, 0, 1), (1, 0, 0, 1)]
+    dispRanges    = [(0, 5)]
+    modRanges     = [(1.25, 3.75), (0, 5), (5, 10)]
+    invModAlphas  = [False, True]
+    width, height = 300, 100
+
+    testcases = it.product(cmaps,
+                           negCmaps,
+                           alphas,
+                           bgColours,
+                           dispRanges,
+                           modRanges,
+                           invModAlphas)
+
+    for testcase in testcases:
+        cmap, negCmap, alpha, bgColour, dispRange, modRange, invModAlpha = testcase
+
+        bmp = cbarbmp.colourBarBitmap(cmap,
+                                      width,
+                                      height,
+                                      negCmap=negCmap,
+                                      alpha=alpha,
+                                      bgColour=bgColour,
+                                      orientation='horizontal',
+                                      modAlpha=True,
+                                      invModAlpha=invModAlpha,
+                                      displayRange=dispRange,
+                                      modRange=modRange)
+
+        fname = gen_file_id('modAlpha', *testcase)
+        fname = f'{fname}.png'
         assert _compare(bmp, fname)